requirements.txt
Mako==1.3.5
requests==2.32.3
retrying==1.3.4
cxxheaderparser==1.3.4
PyYAML>=6.0
